Catch ArgumentOutOfRangeException on invalid dates on OS
In theory, it is possible that a file has set a last access date that is
outside the range of FromUnixTimeSeconds. In this case, an unexpected
ArgumentOutOfRangeException is thrown. This can lead to unexpected
exceptions. For example, FileSystemInfo.LastAccessTime only throws an
ArgumentOutOfRangeException when the caller attempts to set an invalid
access time. But if an invalid access time is already set by the OS, the
caller will still receive such exception. By handling this exception, we
avoid such behaviour.